[LC-609] Apply reviews
 - Replace multiple if evaluations to elif
 - Remove redundant inline if
 - Fix test in tx_cache
 - Add type hints at TC signature
 - Change file names (original test -> test_transactions.py, new
functional test -> test_transaction.py)
 - ... and Refactorings
