+++ testcase/unittest/blockchain/transactions/test_transaction_builder.py
@@ -6,12 +6,13 @@ from loopchain.blockchain.transactions import TransactionBuilder
 from loopchain.blockchain.transactions import genesis, v2, v3
 from loopchain.blockchain.types import Hash32, Signature, ExternalAddress
 from loopchain.crypto.signature import Signer
+from testcase.unittest.blockchain.conftest import TxFactory, TxBuilderFactory
 
 
 @pytest.mark.parametrize("tx_version", [genesis.version, v2.version, v3.version])
 class TestTransactionBuilderBase:
-    def test_builder_version_check(self, tx_builder_factory, tx_version):
-        tx_builder = tx_builder_factory(tx_version=tx_version)
+    def test_builder_version_check(self, tx_builder_factory: TxBuilderFactory, tx_version):
+        tx_builder = tx_builder_factory(tx_version)
 
         if tx_version == genesis.version:
             assert isinstance(tx_builder, genesis.TransactionBuilder)
@@ -20,13 +21,13 @@ class TestTransactionBuilderBase:
         elif tx_version == v3.version:
             assert isinstance(tx_builder, v3.TransactionBuilder)
 
-    def test_reset_cache_resets_members(self, tx_builder_factory, tx_version):
-        tx_builder = tx_builder_factory(tx_version=tx_version)
-        tx_builder.from_address = tx_builder.from_address if tx_builder.from_address else "aa"
-        tx_builder.hash = tx_builder.hash if tx_builder.hash else "aa"
-        tx_builder.signature = tx_builder.signature if tx_builder.signature else "aa"
-        tx_builder.origin_data = tx_builder.origin_data if tx_builder.origin_data else "aa"
-        tx_builder.raw_data = tx_builder.raw_data if tx_builder.raw_data else ""
+    def test_reset_cache_resets_members(self, tx_builder_factory: TxBuilderFactory, tx_version):
+        tx_builder = tx_builder_factory(tx_version)
+        tx_builder.from_address = tx_builder.from_address or "aa"
+        tx_builder.hash = tx_builder.hash or "aa"
+        tx_builder.signature = tx_builder.signature or "aa"
+        tx_builder.origin_data = tx_builder.origin_data or "aa"
+        tx_builder.raw_data = tx_builder.raw_data or ""
 
         tx_builder.reset_cache()
         assert not tx_builder.from_address
@@ -42,8 +43,8 @@ class TestTransactionBuilderBase:
         elif tx_version == v3.version:
             assert not tx_builder._timestamp
 
-    def test_from_address_returns_its_addr_if_exists(self, tx_builder_factory, tx_version):
-        tx_builder = tx_builder_factory(tx_version=tx_version)
+    def test_from_address_returns_its_addr_if_exists(self, tx_builder_factory: TxBuilderFactory, tx_version):
+        tx_builder = tx_builder_factory(tx_version)
         expected_addr = ExternalAddress(os.urandom(ExternalAddress.size))
 
         tx_builder.from_address = expected_addr
@@ -51,16 +52,16 @@ class TestTransactionBuilderBase:
 
         assert expected_addr == built_addr
 
-    def test_from_address_raise_exc_if_no_from_addr_and_no_signer(self, tx_builder_factory, tx_version):
-        tx_builder = tx_builder_factory(tx_version=tx_version)
+    def test_from_address_raise_exc_if_no_from_addr_and_no_signer(self, tx_builder_factory: TxBuilderFactory, tx_version):
+        tx_builder = tx_builder_factory(tx_version)
         tx_builder.reset_cache()
         tx_builder.signer = None
 
         with pytest.raises(RuntimeError):
             assert tx_builder.build_from_address()
 
-    def test_from_address_generate_addr_if_no_from_addr_but_signer(self, tx_builder_factory, tx_version):
-        tx_builder = tx_builder_factory(tx_version=tx_version)
+    def test_from_address_generate_addr_if_no_from_addr_but_signer(self, tx_builder_factory: TxBuilderFactory, tx_version):
+        tx_builder = tx_builder_factory(tx_version)
         tx_builder.reset_cache()
 
         assert not tx_builder.from_address
@@ -68,30 +69,30 @@ class TestTransactionBuilderBase:
 
         assert tx_builder.build_from_address()
 
-    def test_build_origin_data(self, tx_builder_factory, tx_version):
-        tx_builder = tx_builder_factory(tx_version=tx_version)
+    def test_build_origin_data(self, tx_builder_factory: TxBuilderFactory, tx_version):
+        tx_builder = tx_builder_factory(tx_version)
         tx_builder.build_from_address()
         origin_data = tx_builder.build_origin_data()
 
         assert origin_data
 
-    def test_build_hash_returns_valid_hash_form(self, tx_builder_factory, tx_version):
-        tx_builder: TransactionBuilder = tx_builder_factory(tx_version=tx_version)
+    def test_build_hash_returns_valid_hash_form(self, tx_builder_factory: TxBuilderFactory, tx_version):
+        tx_builder: TransactionBuilder = tx_builder_factory(tx_version)
         tx_builder.build_from_address()
         tx_builder.build_origin_data()
         hash_ = tx_builder.build_hash()
 
         assert isinstance(hash_, Hash32)
 
-    def test_build_hash_fails_if_origin_data_is_not_exist(self, tx_builder_factory, tx_version):
-        tx_builder = tx_builder_factory(tx_version=tx_version)
+    def test_build_hash_fails_if_origin_data_is_not_exist(self, tx_builder_factory: TxBuilderFactory, tx_version):
+        tx_builder = tx_builder_factory(tx_version)
         assert tx_builder.origin_data is None
 
         with pytest.raises(RuntimeError, match="origin data is required"):
             assert tx_builder.build_hash()
 
-    def test_sign_builds_signature_and_hash(self, tx_builder_factory, tx_version, mocker):
-        tx_builder: TransactionBuilder = tx_builder_factory(tx_version=tx_version)
+    def test_sign_builds_signature_and_hash(self, tx_builder_factory: TxBuilderFactory, tx_version, mocker):
+        tx_builder: TransactionBuilder = tx_builder_factory(tx_version)
         assert not tx_builder.signature
         assert not tx_builder.hash
 
@@ -103,8 +104,8 @@ class TestTransactionBuilderBase:
         assert isinstance(tx_builder.signature, Signature)
         assert tx_builder.hash
 
-    def test_build_raw_data(self, tx_builder_factory, tx_version):
-        tx_builder = tx_builder_factory(tx_version=tx_version)
+    def test_build_raw_data(self, tx_builder_factory: TxBuilderFactory, tx_version):
+        tx_builder = tx_builder_factory(tx_version)
 
         tx_builder.build_from_address()
         tx_builder.build_origin_data()
@@ -114,8 +115,8 @@ class TestTransactionBuilderBase:
 
         assert raw_data
 
-    def test_build_transaction(self, tx_builder_factory, tx_version):
-        tx_builder = tx_builder_factory(tx_version=tx_version)
+    def test_build_transaction(self, tx_builder_factory: TxBuilderFactory, tx_version):
+        tx_builder = tx_builder_factory(tx_version)
         tx = tx_builder.build()
 
         if tx_version == genesis.version:
@@ -125,8 +126,8 @@ class TestTransactionBuilderBase:
         elif tx_version == v3.version:
             assert isinstance(tx, v3.Transaction)
 
-    def test_sign_transaction(self, tx_builder_factory, tx_version):
-        tx_builder = tx_builder_factory(tx_version=tx_version)
+    def test_sign_transaction(self, tx_builder_factory: TxBuilderFactory, tx_version):
+        tx_builder = tx_builder_factory(tx_version)
         tx = tx_builder.build()
 
         if tx_version == genesis.version:
@@ -135,8 +136,8 @@ class TestTransactionBuilderBase:
         else:
             tx_builder.sign_transaction(tx)
 
-    def test_sign_transaction_raises_exc_if_signer_addr_ne_tx_signer_addr(self, tx_builder_factory, tx_version):
-        tx_builder = tx_builder_factory(tx_version=tx_version)
+    def test_sign_transaction_raises_exc_if_signer_addr_ne_tx_signer_addr(self, tx_builder_factory: TxBuilderFactory, tx_version):
+        tx_builder = tx_builder_factory(tx_version)
         tx = tx_builder.build()
 
         tx_builder.signer = Signer.new()
@@ -150,15 +151,15 @@ class TestTransactionBuilderBase:
 
 
 class TestTransactionBuilder_genesis:
-    def test_build_nid_raises_exc_if_hash_not_exists(self, tx_builder_factory):
-        tx_builder: genesis.TransactionBuilder = tx_builder_factory(tx_version=genesis.version)
+    def test_build_nid_raises_exc_if_hash_not_exists(self, tx_builder_factory: TxBuilderFactory):
+        tx_builder: genesis.TransactionBuilder = tx_builder_factory(genesis.version)
         tx_builder.reset_cache()
 
         with pytest.raises(RuntimeError):
             assert tx_builder.build_nid()
 
-    def test_build_nid_returns_its_nid_if_exists(self, tx_builder_factory):
-        tx_builder: genesis.TransactionBuilder = tx_builder_factory(tx_version=genesis.version)
+    def test_build_nid_returns_its_nid_if_exists(self, tx_builder_factory: TxBuilderFactory):
+        tx_builder: genesis.TransactionBuilder = tx_builder_factory(genesis.version)
         expected_nid = genesis.NID.unknown
         tx_builder.nid = expected_nid
         tx_builder.build_origin_data()
@@ -168,14 +169,14 @@ class TestTransactionBuilder_genesis:
         assert expected_nid == nid_generated
 
     @pytest.mark.parametrize("tx_hash, expected_nid", [
-        (genesis.NTxHash.mainnet, genesis.NID.mainnet),
-        (genesis.NTxHash.testnet, genesis.NID.testnet),
+        ("0x5aa2453a84ba2fb1e3394b9e3471f5dcebc6225fc311a97ca505728153b9d246", genesis.NID.mainnet),
+        ("0x5a7ce1e10a6fd5fb3925a011528f89a5debfead2405f5545a99d1a1310e48c9e", genesis.NID.testnet)
     ])
-    def test_build_nid_with_specific_hash_matches_expected_nid(self, tx_builder_factory, tx_hash: genesis.NTxHash, expected_nid: genesis.NID):
-        tx_builder: genesis.TransactionBuilder = tx_builder_factory(tx_version=genesis.version)
+    def test_build_nid_with_specific_hash_matches_expected_nid(self, tx_builder_factory: TxBuilderFactory, tx_hash, expected_nid: genesis.NID):
+        tx_builder: genesis.TransactionBuilder = tx_builder_factory(genesis.version)
         assert not tx_builder.nid
 
-        tx_builder.hash = tx_hash.value
+        tx_builder.hash = Hash32.fromhex(tx_hash)
         nid_generated = tx_builder.build_nid()
 
         assert nid_generated == expected_nid
