+++ testcase/unittest/blockchain/transactions/test_transaction_serializer.py
@@ -4,14 +4,15 @@ from freezegun import freeze_time
 from loopchain.blockchain.transactions import TransactionSerializer, TransactionVersioner, Transaction
 from loopchain.blockchain.transactions import genesis, v2, v3, v3_issue
 from loopchain.blockchain.types import Hash32
+from testcase.unittest.blockchain.conftest import TxFactory, TxBuilderFactory
 
 tx_versioner = TransactionVersioner()
 
 
 class TestTransactionSerializerBase:
     @pytest.mark.parametrize("tx_version", [genesis.version, v2.version, v3.version])
-    def test_serializer_version_check(self, tx_factory, tx_version):
-        tx: Transaction = tx_factory(tx_version=tx_version)
+    def test_serializer_version_check(self, tx_factory: TxFactory, tx_version):
+        tx: Transaction = tx_factory(tx_version)
         ts = TransactionSerializer.new(version=tx.version, type_=tx.type(), versioner=tx_versioner)
 
         if tx_version == genesis.version:
@@ -29,8 +30,8 @@ class TestTransactionSerializerBase:
         (v2.version, ["from", "to", "value", "fee", "timestamp", "nonce"]),
         (v3.version, ["version", "from", "to", "stepLimit", "timestamp", "nid", "value", "nonce"])
     ])
-    def test_to_origin_data_has_valid_form(self, tx_factory, tx_version, expected_keys):
-        tx: Transaction = tx_factory(tx_version=tx_version)
+    def test_to_origin_data_has_valid_form(self, tx_factory: TxFactory, tx_version, expected_keys):
+        tx: Transaction = tx_factory(tx_version)
         ts = TransactionSerializer.new(version=tx.version, type_=tx.type(), versioner=tx_versioner)
 
         origin_data = ts.to_origin_data(tx)
@@ -41,32 +42,32 @@ class TestTransactionSerializer_genesis:
     tx_version = genesis.version
 
     # TODO: What is diffrence between tx.raw_data and its dict casted?
-    def test_to_raw_data_equals_tx_raw_data(self, tx_factory):
-        tx: Transaction = tx_factory(tx_version=self.tx_version)
+    def test_to_raw_data_equals_tx_raw_data(self, tx_factory: TxFactory):
+        tx: Transaction = tx_factory(self.tx_version)
         ts = TransactionSerializer.new(version=tx.version, type_=tx.type(), versioner=tx_versioner)
 
         raw_data = ts.to_raw_data(tx)
         assert raw_data == tx.raw_data
 
-    def test_to_db_data_equals_tx_raw_data(self, tx_factory):
-        tx: Transaction = tx_factory(tx_version=self.tx_version)
+    def test_to_db_data_equals_tx_raw_data(self, tx_factory: TxFactory):
+        tx: Transaction = tx_factory(self.tx_version)
         ts = TransactionSerializer.new(version=tx.version, type_=tx.type(), versioner=tx_versioner)
 
         db_data = ts.to_db_data(tx)
         assert db_data == tx.raw_data
 
-    def test_to_full_data_equals_tx_raw_data(self, tx_factory):
-        tx: Transaction = tx_factory(tx_version=self.tx_version)
+    def test_to_full_data_equals_tx_raw_data(self, tx_factory: TxFactory):
+        tx: Transaction = tx_factory(self.tx_version)
         ts = TransactionSerializer.new(version=tx.version, type_=tx.type(), versioner=tx_versioner)
 
         full_data = ts.to_full_data(tx)
         assert full_data == tx.raw_data
 
     @pytest.mark.xfail(reason="Check `is_signing` flag at build().")
-    def test_orig_tx_equals_deserialized_tx(self, tx_builder_factory):
+    def test_orig_tx_equals_deserialized_tx(self, tx_builder_factory: TxBuilderFactory):
         with freeze_time():
             # TODO: origin data contains signature, so it affects tx_hash of deserialized tx.
-            tx: genesis.Transaction = tx_builder_factory(tx_version=self.tx_version)\
+            tx: genesis.Transaction = tx_builder_factory(self.tx_version)\
                 .build(is_signing=False)
             ts: genesis.TransactionSerializer = TransactionSerializer.new(version=tx.version, type_=tx.type(), versioner=tx_versioner)
 
@@ -75,8 +76,8 @@ class TestTransactionSerializer_genesis:
         assert tx == tx_restored
 
     @pytest.mark.xfail(raises=KeyError, reason="Genesis tx raw_data has no `tx_hash`?")
-    def test_get_hash(self, tx_factory):
-        tx: Transaction = tx_factory(tx_version=self.tx_version)
+    def test_get_hash(self, tx_factory: TxFactory):
+        tx: Transaction = tx_factory(self.tx_version)
         ts = TransactionSerializer.new(version=tx.version, type_=tx.type(), versioner=tx_versioner)
 
         full_data = ts.to_full_data(tx)
@@ -100,8 +101,8 @@ class TestTransactionSerializer_v2:
     """
     tx_version = v2.version
 
-    def test_to_raw_data_equals_dict_tx_raw_data(self, tx_factory):
-        tx: Transaction = tx_factory(tx_version=self.tx_version)
+    def test_to_raw_data_equals_dict_tx_raw_data(self, tx_factory: TxFactory):
+        tx: Transaction = tx_factory(self.tx_version)
         ts = TransactionSerializer.new(version=tx.version, type_=tx.type(), versioner=tx_versioner)
 
         raw_data = ts.to_raw_data(tx)
@@ -109,8 +110,8 @@ class TestTransactionSerializer_v2:
         assert raw_data == tx.raw_data
         assert tx.raw_data == dict(tx.raw_data)
 
-    def test_to_db_data_equals_full_data(self, tx_factory):
-        tx: Transaction = tx_factory(tx_version=self.tx_version)
+    def test_to_db_data_equals_full_data(self, tx_factory: TxFactory):
+        tx: Transaction = tx_factory(self.tx_version)
         ts = TransactionSerializer.new(version=tx.version, type_=tx.type(), versioner=tx_versioner)
 
         db_data = ts.to_db_data(tx)
@@ -118,8 +119,8 @@ class TestTransactionSerializer_v2:
 
         assert db_data == full_data
 
-    def test_to_full_data_equals_raw_data_with_method(self, tx_factory):
-        tx: Transaction = tx_factory(tx_version=self.tx_version)
+    def test_to_full_data_equals_raw_data_with_method(self, tx_factory: TxFactory):
+        tx: Transaction = tx_factory(self.tx_version)
         ts = TransactionSerializer.new(version=tx.version, type_=tx.type(), versioner=tx_versioner)
 
         full_data = ts.to_full_data(tx)
@@ -129,8 +130,8 @@ class TestTransactionSerializer_v2:
         assert full_data == raw_data
 
     @pytest.mark.xfail(reason="Ignore `to_address (MalformedStr)` attrs ?")
-    def test_orig_tx_equals_deserialized_tx(self, tx_factory):
-        tx: Transaction = tx_factory(tx_version=self.tx_version)
+    def test_orig_tx_equals_deserialized_tx(self, tx_factory: TxFactory):
+        tx: Transaction = tx_factory(self.tx_version)
         ts = TransactionSerializer.new(version=tx.version, type_=tx.type(), versioner=tx_versioner)
 
         tx_restored = ts.from_(tx.raw_data)
@@ -140,8 +141,8 @@ class TestTransactionSerializer_v2:
 
         assert tx == tx_restored
 
-    def test_get_hash(self, tx_factory):
-        tx: Transaction = tx_factory(tx_version=self.tx_version)
+    def test_get_hash(self, tx_factory: TxFactory):
+        tx: Transaction = tx_factory(self.tx_version)
         ts = TransactionSerializer.new(version=tx.version, type_=tx.type(), versioner=tx_versioner)
 
         tx_hash = ts.get_hash(tx.raw_data)
@@ -166,8 +167,8 @@ class TestTransactionSerializer_v3:
 
     tx_version = v3.version
 
-    def test_to_raw_data_equals_dict_tx_raw_data(self, tx_factory):
-        tx: Transaction = tx_factory(tx_version=self.tx_version)
+    def test_to_raw_data_equals_dict_tx_raw_data(self, tx_factory: TxFactory):
+        tx: Transaction = tx_factory(self.tx_version)
         ts = TransactionSerializer.new(version=tx.version, type_=tx.type(), versioner=tx_versioner)
 
         raw_data = ts.to_raw_data(tx)
@@ -175,8 +176,8 @@ class TestTransactionSerializer_v3:
         assert raw_data == tx.raw_data
         assert tx.raw_data == dict(tx.raw_data)
 
-    def test_to_db_data_equals_dict_raw_data(self, tx_factory):
-        tx: Transaction = tx_factory(tx_version=self.tx_version)
+    def test_to_db_data_equals_dict_raw_data(self, tx_factory: TxFactory):
+        tx: Transaction = tx_factory(self.tx_version)
         ts = TransactionSerializer.new(version=tx.version, type_=tx.type(), versioner=tx_versioner)
 
         db_data = ts.to_db_data(tx)
@@ -184,8 +185,8 @@ class TestTransactionSerializer_v3:
         assert db_data == tx.raw_data
         assert tx.raw_data == dict(tx.raw_data)
 
-    def test_to_full_data_equals_to_db_data_with_tx_hash(self, tx_factory):
-        tx: Transaction = tx_factory(tx_version=self.tx_version)
+    def test_to_full_data_equals_to_db_data_with_tx_hash(self, tx_factory: TxFactory):
+        tx: Transaction = tx_factory(self.tx_version)
         ts = TransactionSerializer.new(version=tx.version, type_=tx.type(), versioner=tx_versioner)
 
         full_data = ts.to_full_data(tx)
@@ -194,8 +195,8 @@ class TestTransactionSerializer_v3:
 
         assert db_data == full_data
 
-    def test_orig_tx_equals_deserialized_tx(self, tx_factory):
-        tx: Transaction = tx_factory(tx_version=self.tx_version)
+    def test_orig_tx_equals_deserialized_tx(self, tx_factory: TxFactory):
+        tx: Transaction = tx_factory(self.tx_version)
         ts = TransactionSerializer.new(version=tx.version, type_=tx.type(), versioner=tx_versioner)
 
         full_data = ts.to_full_data(tx)
@@ -203,8 +204,8 @@ class TestTransactionSerializer_v3:
 
         assert tx == tx_restored
 
-    def test_get_hash(self, tx_factory):
-        tx: Transaction = tx_factory(tx_version=self.tx_version)
+    def test_get_hash(self, tx_factory: TxFactory):
+        tx: Transaction = tx_factory(self.tx_version)
         ts = TransactionSerializer.new(version=tx.version, type_=tx.type(), versioner=tx_versioner)
 
         full_data = ts.to_full_data(tx)
@@ -217,14 +218,14 @@ class TestTransactionSerializer_v3_issue:
     tx_version = v3_issue.version
     type_ = "base"
 
-    def test_serializer_version_check(self, tx_factory):
-        tx: Transaction = tx_factory(tx_version=self.tx_version)
+    def test_serializer_version_check(self, tx_factory: TxFactory):
+        tx: Transaction = tx_factory(self.tx_version)
         ts = TransactionSerializer.new(version=tx.version, type_=self.type_, versioner=tx_versioner)
 
         assert isinstance(ts, v3_issue.TransactionSerializer)
 
-    def test_to_origin_data_has_valid_form(self, tx_factory):
-        tx: Transaction = tx_factory(tx_version=self.tx_version)
+    def test_to_origin_data_has_valid_form(self, tx_factory: TxFactory):
+        tx: Transaction = tx_factory(self.tx_version)
         ts = TransactionSerializer.new(version=tx.version, type_=self.type_, versioner=tx_versioner)
 
         origin_data = ts.to_origin_data(tx)
@@ -232,8 +233,8 @@ class TestTransactionSerializer_v3_issue:
 
         assert set(origin_data) == set(expected_keys)
 
-    def test_to_raw_data_equals_dict_tx_raw_data(self, tx_factory):
-        tx: Transaction = tx_factory(tx_version=self.tx_version)
+    def test_to_raw_data_equals_dict_tx_raw_data(self, tx_factory: TxFactory):
+        tx: Transaction = tx_factory(self.tx_version)
         ts = TransactionSerializer.new(version=tx.version, type_=self.type_, versioner=tx_versioner)
 
         raw_data = ts.to_raw_data(tx)
@@ -241,8 +242,8 @@ class TestTransactionSerializer_v3_issue:
         assert raw_data == tx.raw_data
         assert tx.raw_data == dict(tx.raw_data)
 
-    def test_to_db_data_equals_dict_raw_data(self, tx_factory):
-        tx: Transaction = tx_factory(tx_version=self.tx_version)
+    def test_to_db_data_equals_dict_raw_data(self, tx_factory: TxFactory):
+        tx: Transaction = tx_factory(self.tx_version)
         ts = TransactionSerializer.new(version=tx.version, type_=self.type_, versioner=tx_versioner)
 
         db_data = ts.to_db_data(tx)
@@ -250,8 +251,8 @@ class TestTransactionSerializer_v3_issue:
         assert db_data == tx.raw_data
         assert tx.raw_data == dict(tx.raw_data)
 
-    def test_to_full_data_equals_to_db_data_with_tx_hash(self, tx_factory):
-        tx: Transaction = tx_factory(tx_version=self.tx_version)
+    def test_to_full_data_equals_to_db_data_with_tx_hash(self, tx_factory: TxFactory):
+        tx: Transaction = tx_factory(self.tx_version)
         ts = TransactionSerializer.new(version=tx.version, type_=self.type_, versioner=tx_versioner)
 
         full_data = ts.to_full_data(tx)
@@ -261,9 +262,9 @@ class TestTransactionSerializer_v3_issue:
         assert db_data == full_data
 
     @pytest.mark.xfail(reason="How to test?")
-    def test_orig_tx_equals_deserialized_tx(self, tx_factory):
+    def test_orig_tx_equals_deserialized_tx(self, tx_factory: TxFactory):
         with freeze_time():
-            tx: Transaction = tx_factory(tx_version=self.tx_version)
+            tx: Transaction = tx_factory(self.tx_version)
             ts = TransactionSerializer.new(version=tx.version, type_=self.type_, versioner=tx_versioner)
 
         full_data = ts.to_full_data(tx)
@@ -271,8 +272,8 @@ class TestTransactionSerializer_v3_issue:
 
         assert tx == tx_restored
 
-    def test_get_hash(self, tx_factory):
-        tx: Transaction = tx_factory(tx_version=self.tx_version)
+    def test_get_hash(self, tx_factory: TxFactory):
+        tx: Transaction = tx_factory(self.tx_version)
         ts = TransactionSerializer.new(version=tx.version, type_=self.type_, versioner=tx_versioner)
 
         full_data = ts.to_full_data(tx)
