+++ testcase/unittest/blockchain/transactions/test_transaction_verifier_cache.py
@@ -1,16 +1,12 @@
-import os
-import random
-
 import pytest
 
 from loopchain.blockchain.exception import TransactionInvalidHashError
 from loopchain.blockchain.exception import TransactionInvalidSignatureError
-from loopchain.blockchain.transactions import TransactionBuilder, TransactionVerifier, TransactionVersioner
-from loopchain.blockchain.transactions import genesis, v2, v3, v3_issue
-from loopchain.blockchain.types import ExternalAddress
+from loopchain.blockchain.transactions import TransactionVerifier, TransactionVersioner
+from loopchain.blockchain.transactions import v2, v3
 from loopchain.blockchain.types import Hash32
 from loopchain.blockchain.types import Signature
-from loopchain.crypto.signature import Signer
+from testcase.unittest.blockchain.conftest import TxFactory
 
 tx_versioner = TransactionVersioner()
 
@@ -22,10 +18,10 @@ class TestTxCache:
     target_attrs = ["hash", "signature"]
 
     @pytest.mark.parametrize("target_attr", target_attrs)
-    def test_no_cache_attr_in_tx_if_not_verified(self, tx_version, tx_factory, target_attr):
+    def test_no_cache_attr_in_tx_if_not_verified(self, tx_version, tx_factory: TxFactory, target_attr):
         """Check that Transaction has no attribute until the verification func has been invoked"""
-        tx = tx_factory(tx_version=tx_version)
-        if "cache" in dir(tx):
+        tx = tx_factory(tx_version)
+        if any("cache" in attr for attr in dir(tx)):
             raise AttributeError("Something wrong. Transaction has cached value when initialized.")
 
         with pytest.raises(AttributeError):
@@ -33,9 +29,9 @@ class TestTxCache:
 
     @pytest.mark.parametrize("raise_exception", [True, False])
     @pytest.mark.parametrize("target_attr", ["hash", "signature"])
-    def test_has_cache_attr_when_verified_successfully(self, tx_version, tx_factory, raise_exception, target_attr):
+    def test_has_cache_attr_when_verified_successfully(self, tx_version, tx_factory: TxFactory, raise_exception, target_attr):
         """Check that the verification result has been cached successfully"""
-        tx = tx_factory(tx_version=tx_version)
+        tx = tx_factory(tx_version)
         tv = TransactionVerifier.new(tx.version, tx.type(), tx_versioner, raise_exceptions=raise_exception)
 
         verify_func_name = f"verify_{target_attr}"
@@ -49,9 +45,9 @@ class TestTxCache:
         (TransactionInvalidHashError, "hash", Hash32.new()),
         (TransactionInvalidSignatureError, "signature", Signature.new()),
     ])
-    def test_exception_cached_when_raised_exception_while_verification(self, tx_version, tx_factory, raise_exception, expected_exc, target_attr, fake_value, monkeypatch):
+    def test_exception_cached_when_raised_exception_while_verification(self, tx_version, tx_factory: TxFactory, raise_exception, expected_exc, target_attr, fake_value, monkeypatch):
         """Check that the exception successfully cached when raised any exception while verification step"""
-        tx = tx_factory(tx_version=tx_version)
+        tx = tx_factory(tx_version)
         tv = TransactionVerifier.new(tx.version,  tx.type(), tx_versioner, raise_exceptions=raise_exception)
         verify_func = getattr(tv, f"verify_{target_attr}")
 
@@ -78,13 +74,13 @@ class TestTxCache:
         (TransactionInvalidHashError, "hash", Hash32.new()),
         (TransactionInvalidSignatureError, "signature", Signature.new()),
     ])
-    def test_verify_success_and_no_exc_with_fake_value_at_second(self, tx_version, tx_factory, raise_exception, expected_exc, target_attr, fake_value, monkeypatch):
+    def test_verify_success_and_no_exc_with_fake_value_at_second(self, tx_version, tx_factory: TxFactory, raise_exception, expected_exc, target_attr, fake_value, monkeypatch):
         """Check that the result is successfully cached and bypasses further verifications which could raise exceptions.
 
         Do not apply this usecase in code!
         This test aims the reliablity of cache logic, not for the usefulness of this case.
         """
-        tx = tx_factory(tx_version=tx_version)
+        tx = tx_factory(tx_version)
         tv = TransactionVerifier.new(tx.version,  tx.type(), tx_versioner, raise_exceptions=raise_exception)
         verify_func = getattr(tv, f"verify_{target_attr}")
 
@@ -106,9 +102,9 @@ class TestTxCache:
 
     @pytest.mark.parametrize("tag", ["before_cache", "after_cache"])
     @pytest.mark.parametrize("target_attr", ["hash", "signature"])
-    def test_benchmark_verify(self, benchmark, tx_version, tx_factory, target_attr, tag):
+    def test_benchmark_verify(self, benchmark, tx_version, tx_factory: TxFactory, target_attr, tag):
         """Benchmark the elapsed time of verification func in various cases."""
-        tx = tx_factory(tx_version=tx_version)
+        tx = tx_factory(tx_version)
         tv = TransactionVerifier.new(tx.version,  tx.type(), tx_versioner)
         verify_func = getattr(tv, f"verify_{target_attr}")
